Add expired unlock token scenario
