+++ cypress/integration/common/email.js
@@ -6,6 +6,7 @@ import AddUserPage from '../../pages/add_user_page'
 import EditUserPage from '../../pages/edit_user_page'
 import LastEmailPage from '../../pages/last_email_page'
 import MainMenu from '../../pages/menus/main_menu'
+import ResendUnlockPage from '../../pages/resend_unlock_page'
 import SignInPage from '../../pages/sign_in_page'
 import UsersPage from '../../pages/users_page'
 
@@ -101,6 +102,17 @@ When('I follow the link to unlock my account', () => {
   })
 })
 
+And('request another unlock email', () => {
+  SignInPage.visit()
+  SignInPage.resendUnlockLink().click()
+  ResendUnlockPage.confirm()
+
+  cy.get('@user').then((user) => {
+    ResendUnlockPage.email().type(user.email)
+    ResendUnlockPage.resendUnlockInstructions().click()
+  })
+})
+
 Then('I will see confirmation my account is unlocked', () => {
   cy.get('.col > .alert')
     .should(
@@ -118,6 +130,18 @@ But('I miss the first invitation email', () => {
 
       cy.wrap(link).as('firstLink')
       cy.log(`The first link is ${link}`)
+    })
+  })
+})
+
+But('I miss the first unlock email', () => {
+  cy.get('@user').then((user) => {
+    LastEmailPage.lastEmail([user.email, 'Your account has been locked'])
+
+    cy.get('@lastEmail').then((lastEmail) => {
+      const link = LastEmailPage.extractUnlockAccountLink(lastEmail.last_email.body)
+
+      cy.wrap(link).as('firstLink')
     })
   })
 })
@@ -147,5 +171,13 @@ Then('the TCM will confirm the user has been reinvited', () => {
     .should(
       'contain.text',
       'User reinvited'
+    )
+})
+
+Then('I will see confirmation an unlock email has been sent', () => {
+  cy.get('.col > .alert')
+    .should(
+      'contain.text',
+      'If your account exists, you will receive an email with instructions for how to unlock it in a few minutes.'
     )
 })
