+++ cypress/integration/email/expired_tokens/expired_tokens.js
@@ -1,17 +1,21 @@
 import { And, Then } from 'cypress-cucumber-preprocessor/steps'
 
 import MainMenu from '../../../pages/menus/main_menu'
+import ResendUnlockPage from '../../../pages/resend_unlock_page'
 
-And('I try to accept the first invitation', () => {
+And('I try to accept the first invitation email', () => {
   MainMenu.user.getOption('Sign out').click()
   cy.get('h1').should('contain', 'Sign in')
   cy.url().should('include', '/auth/sign_in')
 
-  cy.get('@user').then((user) => {
-    cy.get('@firstLink').then((firstLink) => {
-      cy.visit(firstLink).then(() => {
-      })
-    })
+  cy.get('@firstLink').then((firstLink) => {
+    cy.visit(firstLink)
+  })
+})
+
+And('I try to accept the first unlock email', () => {
+  cy.get('@firstLink').then((firstLink) => {
+    cy.visit(firstLink)
   })
 })
 
@@ -25,4 +29,9 @@ Then('the TCM will confuse me and not be helpful', () => {
   cy.url().should('include', '/auth/sign_in')
 })
 
-// You need to sign in before continuing.
+Then('I will be directed to the resend unlock page and told my token is invalid', () => {
+  ResendUnlockPage.confirm()
+
+  cy.get('h2').should('contain', 'error prohibited this user from being saved')
+  cy.get('div#error_explanation > ul > li').should('contain', 'Unlock token is invalid')
+})
